+++ src/collections/mod.rs
@@ -1,0 +1,5 @@
+mod dequeue;
+mod vec;
+
+pub use dequeue::DynamicDeque;
+pub use vec::DynamicVec;
