+++ src/lib.rs
@@ -10,6 +10,7 @@ mod worker;
 
 pub use backend::*;
 pub use bus::*;
+pub use collections::*;
 pub use controller::Controller;
 pub use owner::Owner;
 pub use task::{NetIncoming, NetOutgoing, Task, TaskGroup, TaskGroupOutput, TaskInput, TaskOutput};
